Hardware class for Rev kit Robot (RevBotHardware).
The class defines a Rev kit robot (singleton bc there is only one robot). It has all the hardware components such as motors and servos for the different systems, it has an initialization method for initializing all the necessary components, and it has get and set methods to use the components in other classes.